Restructure evil wallet package
