+++ tools/evil-spammer/interactive/menu.go
@@ -1,4 +1,4 @@
-package main
+package interactive
 
 import (
 	"fmt"
