+++ tools/evil-spammer/programs/quick-test.go
@@ -1,4 +1,4 @@
-package main
+package programs
 
 import (
 	"time"
